Move syncExternalService into the backend package
